requirements_render.txt
Django==5.0.6
django-crispy-forms==2.1
python-dotenv==1.0.0
gunicorn==21.2.0